+++ app/Http/Controllers/MaedaoshiController.php
@@ -149,11 +149,11 @@ class MaedaoshiController extends Controller
 
 		$query = "select material_number, sum(plan) as plan, sum(actual) as actual from
 		(
-			select material_number, quantity as plan, 0 as actual from production_schedules where due_date >= '".$first."' and due_date <= '".$last."'
+		select material_number, quantity as plan, 0 as actual from production_schedules where due_date >= '".$first."' and due_date <= '".$last."'
 
-			union all
+		union all
 
-			select material_number, 0 as plan, quantity as actual from flo_details where date(created_at) >= '".$first."' and date(created_at) <= '".$last."'
+		select material_number, 0 as plan, quantity as actual from flo_details where date(created_at) >= '".$first."' and date(created_at) <= '".$last."'
 		) as result
 		group by result.material_number
 		having plan <= actual and material_number = '".$request->get('material')."'";
@@ -615,21 +615,39 @@ class MaedaoshiController extends Controller
 					$code_generator_pd->save();
 				}
 
-				$log_process = LogProcess::firstOrNew([
-					'process_code' => '5',
-					'serial_number' => $serial_number,
-					'model' => $material->model,
-					'manpower' => 2,
-					'quantity' => $actual,
-					'created_by' => $id
-				]);
-				$log_process->save();
+				// $log_process = LogProcess::firstOrNew([
+				// 	'process_code' => '5',
+				// 	'origin_group_code' => $material->origin_group_code,
+				// 	'serial_number' => $serial_number,
+				// 	'model' => $material->model,
+				// 	'manpower' => 2,
+				// 	'quantity' => 1,
+				// 	'created_by' => $id
+				// ]);
+				// $log_process->save();
 
-				$inventory_stamp = StampInventory::where('serial_number', '=', $serial_number)
-				->where('model', '=', $material->model)
-				->first();
-				if($inventory_stamp != null){
-					$inventory_stamp->forceDelete();
+				if($material->origin_group_code == '041'){
+					$log_process = LogProcess::updateOrCreate(
+						[
+							'process_code' => '5', 
+							'serial_number' => $serial_number,
+							'origin_group_code' => $material->origin_group_code
+						],
+						[
+							'model' => $material->model,
+							'manpower' => 2,
+							'quantity' => $actual,
+							'created_by' => $id,
+							'created_at' => date('Y-m-d H:i:s')
+						]
+					);
+
+					$inventory_stamp = StampInventory::where('serial_number', '=', $serial_number)
+					->where('origin_group_code', '=', $material->origin_group_code)
+					->first();
+					if($inventory_stamp != null){
+						$inventory_stamp->forceDelete();
+					}
 				}
 
 				$response = array(
@@ -710,22 +728,40 @@ class MaedaoshiController extends Controller
 				$printer->cut();
 				$printer->close();
 
-				$log_process = LogProcess::firstOrNew([
-					'process_code' => '5',
-					'serial_number' => $serial_number,
-					'model' => $material->model,
-					'manpower' => 2,
-					'quantity' => $actual,
-					'created_by' => $id
-				]);
-				$log_process->save();
+				// $log_process = LogProcess::firstOrNew([
+				// 	'process_code' => '5',
+				// 	'origin_group_code' => $material->origin_group_code,
+				// 	'serial_number' => $serial_number,
+				// 	'model' => $material->model,
+				// 	'manpower' => 2,
+				// 	'quantity' => 1,
+				// 	'created_by' => $id
+				// ]);
+				// $log_process->save();
 
-				$inventory_stamp = StampInventory::where('serial_number', '=', $serial_number)
-				->where('model', '=', $material->model)
-				->first();
-				if($inventory_stamp != null){
-					$inventory_stamp->forceDelete();
-				}
+				if($material->origin_group_code == '041'){
+					$log_process = LogProcess::updateOrCreate(
+						[
+							'process_code' => '5', 
+							'serial_number' => $serial_number,
+							'origin_group_code' => $material->origin_group_code
+						],
+						[
+							'model' => $material->model,
+							'manpower' => 2,
+							'quantity' => $actual,
+							'created_by' => $id,
+							'created_at' => date('Y-m-d H:i:s')
+						]
+					);
+
+					$inventory_stamp = StampInventory::where('serial_number', '=', $serial_number)
+					->where('origin_group_code', '=', $material->origin_group_code)
+					->first();
+					if($inventory_stamp != null){
+						$inventory_stamp->forceDelete();
+					}
+				}				
 
 				$response = array(
 					'status' => true,
